Implement StreamApi with -S
