+++ .eslintrc.json
@@ -1,10 +1,13 @@
 {
   "extends": "eslint:recommended",
   "parserOptions": {
-    "ecmaVersion": 2017
+    "ecmaVersion": 2020
   },
   "env": {
-    "es6": true,
+    "es2020": true,
     "node": true
+  },
+  "rules": {
+    "no-console": "off"
   }
 }
